Make constant initialization failure an unrecoverable error
Summary: With the introduction of enum classes, we relaxed a long held restriction that objects may not be class constants. Of course, object initialization can fail with exceptions. If this were to happen when the object being initialized is a class constant (that is, an enum class member) there is no scenario in which it makes sense for the program not to terminate. Thus, we catch such exceptions in `clsCnsGet()` and `raise_error()`.

Reviewed By: paulbiss

Differential Revision: D27300014

fbshipit-source-id: cd1abf447fe592148d8932c39081789c8a2d8091