--dump-symbol-info uses FindUtils.file_filter
Summary:
What does "hh --dump-symbol-info" do when you give it an inappropriate filename?
1. If the filename doesn't exist, then ClientCheck.expand_path reports "file not found" and exits
2. If the filename exists but isn't within the project root, then it fails with an assert ""%s is not a prefix of %s" and goes into an infinite retry loop.
3. If the filename does exist within the project root but doesn't have FileInfo.t associated with it in the naming table, then it's silently skipped
4. Otherwise it dumps symbols from the file.

***What should it do?***

It struck me as a code smell that we're testing for the presence of FileInfo.t but without actually doing anything with that FileInfo.t.

Q1. What kind of file in the repository wouldn't be in the naming-table anyway? The only one I can think of is a file that doesn't pass FindUtils.file_filter.

Q2. What's the intended behavior of --dump-symbol-info for such a file? I think it's reasonable to skip it.

I'm working through all places that consume FileInfo.t from the naming table to learn why they do it and what they retrieve -- do they need strings? or name hashes?

In this case, I think it was an abuse to fetch a FileInfo.t from the naming-table just to decide whether or not to dump symbols from the file. I think a more appropriate behavior would be to use FindUtils.file_filter.

Reviewed By: CatherineGasnier

Differential Revision: D27136048

fbshipit-source-id: 123fbe66e20645c7f69a5534b8a540b55fae04a7