Naming_provider.remove_batch should take list, not set
Summary:
REVIEWER NOTE: (1) start with naming_provider.mli, (2) then read the lengthy comment in naming_provider.ml, (3) then read naming_heap.mli, (4) then read the change in naming_heap.ml which converts list into sets at point of need, (5) then the rest of the diff is mechanical changes.

Why does Naming_provider.remove_batch take a SSet?

Answer 1: because it wraps SharedMem.remove_batch, which itself takes a Set.

Answer 2: to embody the semantics that it's a collection of unique (non-repeated) elements

I think that "answer 2" has mislead us to the point of having a bug. It has to do with case sensitivity, and the question of whether those elements are unique with respect to case-sensitive equality or to case-insensitive equality.

1. Imagine the old names in a file was the set {Aa,aA} and the new names is {aa}.
2. Our codebase currently glosses over the definition of equality, defaulting just to use case-sensitive equality. It therefore deems that symbols {Aa,aA} have been removed, and symbol {aa} has been added.
3. When remove_batch is called on the properly-cased reverse naming table, things are fine: the properly-cased naming table did indeed have two separate entries, one for "Aa" and one for "aA", and each gets removed.
4. When remove_batch is called on the lower-cased reverse naming table (confusingly called the "canon table") it doesn't work. (1) it started with only one entry for "aa". (2) To remove Aa, we do that by removing the lower-case version of it. (3) Next, to remove aA, we do that by removing the lower-case version of it. But there's nothing to remove! This will give rise to a failwith in the current codebase.

I think the danger comes because SSet lulls us into a false sense of security about equality. I think we need to be more careful about equality.

In this diff, therefore, I'm leaving it as a string list for a while longer. I only convert from list into string at the final point, right before we get to removing keys from a single heap.

(This has an added clean-code advantage! Every callstite in the codebase stared with a list, and turned it into a set solely for the purpose of calling remove_batch. This diff removes all that cleanup, since conversion to SSet happens in a single common place downstream.)

Why now? I discovered this issue because I'm in the process of switching away from strings, towards using abstract types for symbol names, e.g. FunName (for the name of a function which supports case-sensitive equality only) and IFunName (for the name of a function which supports case-insensitive equality as well). I purposefully made it so you had to explicitly indicate which version of equality you want to use. When I tried to port Naming_provider to use these new abstract types, the compiler alerted me that something fishy was going on. And indeed, upon code inspection, I think it was.

Reviewed By: CatherineGasnier

Differential Revision: D27223013

fbshipit-source-id: 399b4ba442b6870de60c141716074c604f8a781d