Check dynamic member access within dynamic class for private members
Summary:
We impose no restrictions on the types of private members in classes that are marked `<<__SoundDynamicCallable>>`. This is ok *so long as* a member of the same name isn't accessed from code in the class through an object of type `dynamic`. For that, we must check at the use-site.

(1) A read expression `$x->prop` for `$x:dynamic` is rejected if there exists `prop` in the enclosing class whose type is not a subtype of `dynamic` and whose accessibility is private.
(2) A write expression `$x->prop = ...` for `$x:dynamic` is rejected if there exists `prop` in the enclosing class whose type is not enforceable and whose accessibility is private.

A subsequent diff will tackle private methods.

Differential Revision: D27325926

fbshipit-source-id: c058d69525f7873909384276d5487b71513040c4