Clean up, part 3: burn dvarrays out of HHBBC's type system
Summary:
Delete BVArr, BDArr, BVecish, and BDictish, and propagate the results as far as they go.

Lots of code is now dead. Code that used to refer to "vecish" or "dictish" can now use just "vec" or "dict". We drop these types from RepoAuthType, which also necessitates a small cleanup in the runtime.

Reviewed By: jano

Differential Revision: D27330918

fbshipit-source-id: b5cb3d9914a8d61640e90e7116c22969afaa6462