Stop locking test files in run.php
Summary:
I have no idea why we've been doing this. We never run the same test more than once in one test run (if we run in both interp and jit mode, then we do that _inside_ run_test, while holding the lock once). Perhaps it was to protect multiple test runs in the same repo from stepping on each other, but locking should not be needed for that since we ensured that no tests write to the repo but instead use proper tempdirs. If any remaining tests flake when run by multiple runners concurrently, we need to fix those tests, not work around it in the runner.

Stop trying to lock tests in this way. Rename run_and_lock_test to run_and_log_test to better reflect what it now does.

Differential Revision: D27227677

fbshipit-source-id: 921cc0e68793e544c87911a11e9b4a4b59c9bc39