Remove run.php support for --hack-only option.
Summary:
When this was created, the intent was to ignore all tests thate were still in PHP and allowed PHP features that were forbidden in hack. That's not a thing any more, but this flag still exists, until recently the one-bit-refcount test targets still passed it, and for `./hphp/test/run --hack-only all` it causes 1790 tests to skip (though not for any good reason - the vast najority are skipped because they start with `<?hh <<__EntryPoint>> function main(): void {` all on one line (from a codemod), and that doesn't match either of the 2 first lines that the implementation of this feature allows).

Remove support for this recently dead (D27276919) feature.

Differential Revision: D27183516

fbshipit-source-id: c7c8c9d6a91b3adb961d78eb5b03441db9c9fb8b