Add .runif file support to run.php
Summary: Add a mechanism for skipping tests that's similar in spirit to .skipif but instead of allowing one to run arbitrary code it can only skip based on pre-defined reasons understood by the test runner. Add support for the most common reasons to use a .skipif: testing for the OS type, for whether we're running the test as root, and for the presence of an extension, function, class, method, or constant.

Differential Revision: D27325002

fbshipit-source-id: 99ea0b6d02843bc5cadc6dbfeea5c2b87a68c12a