Clean up, part 5: Delete dvarrays in jit::Type
Summary:
Delete TArr, TVArr, and TDArr, and then propagate the change as far as it goes.

Making this change reveals that several IR ops are now dead. We clean those up and simplify type checks throughout the JIT. Here are the dead ops:
- AllocVArray
- NewDArray
- NewStructDArray
- NewBespokeStructDArray
- ConvArrLikeToVArr
- ConvClsMethToVArr
- ConvObjToVArr
- ConvArrLikeToDArr
- ConvClsMethToDArr
- ConvObjToDArr
- CheckMixedArrayOffset
- ElemMixedArrayK
- TagProvenanceHere

Take a look at emitTagProvenanceHere for just about the best emitter a bytecode could have!

There's one change that looks functional in this diff but is mostly silent. In emitIdx, we now punt if the base may be a Hack array but is not a subtype of a specific Hack array. The old code is incorrect, but happens to work in almost all cases because we break tracelets when the base type is not known. (There are ways to trick the old code into producing incorrect behavior, but they're too rare to happen in prod.)

Reviewed By: oulgen

Differential Revision: D27382096

fbshipit-source-id: 06edea1f483172e7a1b418713ee2182f2609354c