Clean up, part 8: delete KindOfDVArray
Summary:
RIP KindOfDVArray 2020-2021

This diff got a bit interesting. It revealed a few HADVAs bugs - see D27388732. The remainder is just DCE, though.
- Delete DataTypes: KindOfVArray, KindOfPersistentVArray, KindOfDArray, KindOfPersistentDArray.
- Delete DataType predicates: all "HAM-safe" ones (HAMSafeVArray is now just isVec, etc.), isArray, isPHPArray, isHackArray (now just isArrayLike)
- Delete AnnotTypes: VArray and DArray (DataType-specific TypeConstraints) and VArrOrDArr
- Delete all PHP-array-associated APCTypedValue kinds.

Note that I did NOT delete the header kinds associated with dvarrays. Doing so triggers a massive cleanup of array logic which quickly got out of hand. It should be a separate diff. That means that toDataType returns an invalid type for dvarrays (of which we have a few left: the static empty ones).

Reviewed By: jano

Differential Revision: D27388657

fbshipit-source-id: 87918ef0a91a3f7e97592a49756f054f79c158e6