builtin-hsl: add experimental/placeholder HSL to fbcode
Summary:
This is just the `.hhconfig` and `Vec\map()` from the github HSL repository, renamed to avoid conflicts until the migration starts. Committing it as a base for build system and test infra.

Using `hphp/hsl` as it's meant to be shared between the typechecker, hhvm, and other projects - it definitely
doesn't belong in `hphp/runtime/ext/`, and if it's in `hphp/system/php` it can't be disabled. It also makes it
clearer that it shouldn't be thought of as just 'part of hhvm'.

When stuff needs HHVM-specific native code, the relevent part should be deleted from this dir and added
to `hphp/runtime/ext`

When the migration starts, this will be replaced with the full HSL, in the correct namespaces.

Reviewed By: periodic1236

Differential Revision: D27174649

fbshipit-source-id: 01ea0080f1edc14abca29ba91a2539bd2056c43f