HackC implementation for getValue()
Summary:
The implementation of getValue() as `return static::getValues()[$atom];` forces evaluation of all of an enum class's members. Instead, I model this as a sort of `static::$atom` if constants could be accessed this way.

The `get_enum_member_by_label` does not exist, and will throw "function not found" if it is somehow invoked. The expression
```
__SystemLib\get_enum_member_by_label($atom)
```
will emit
```
LateBoundCls    # gets class referred to by static::
ClsCnsL $atom   # this is the bytecode used for HH\MemberOf in the body

Reviewed By: vsiles

Differential Revision: D27281792

fbshipit-source-id: 7d69516a750f13c73a85a0b5f08d2827483a7547