Revise semantics of Readonly immediate
Summary:
After discussing with James, I think it makes more sense to revise the semantics of the readonly immediate. Previously, the semantics related to lvals/rvals, but this notion degrades when you deal with the readonly immediate on member ops as well as the setop flavored bytecodes.

I have revised to the semantics of the readonly immediate to be
```
CGetS Mutable means the static property result of this bytecode must be Mutable
SetS ReadOnly means the static property written to by this bytecode must be ReadOnly
x Any means that this bytecode will not error as we do not care
```

HackC knows whether we need a "must be mutable" vs "must be readonly" check so we should move this logic forward into HackC, and simply execute this check in HHVM. It also makes the bytecode easier to read and reason about.

> x Any means that this bytecode will not error as we do not care

Note: this means the default should in fact be ReadOnlyOp::Any

Note: HackC will not know about whether we need a "must be mutable" or "must be readonly" check for when we decide to allow COW types to be modified. This is ok. Once we determine how we want to track that information in hhvm, we can still emit a "Must be Mutable" check from HackC and make an exception for "readonly but value type." in hhvm.

I wrote up something short about the semantics of the immediates as well as some thoughts on implementation details in a rough quip [here](https://fb.quip.com/mrakAVlV5JMN?range_rows=4&range_cols=5#s:WMPACAWlQZM;WMPACAA8PLk).

Reviewed By: paulbiss

Differential Revision: D27097944

fbshipit-source-id: 897d29e10a43fe822702d93f213bb476db8b9af3