Rollback TransLocMarker on translation overflow and protect it's writes
Summary:
TransLocMarker actually reserves space in the codecache for writing the sizes
of the cold and frozen section into the TC.  This means it is possible for
`startMarker` to overflow the TC.  Move `startMarker` into the try catch block
for TC overflow detection.  Also when we do overflow rollback the current
marker to possibly make space for other regions that might be tranlsated next.
(or to prevent leaving an unecessary gap in the TC).

I made rollback return a TransRange struct representing the range being rolled
back.  This is used for bytes written detection since markEnd (which produces a
TransRange), also writes to the dword markStart saved.  Since there may have
not been enough TC space for that dword, and reserving it may be the reason for
overflow, we cannot call `markEnd`.

Finally AsyncFuncs used in job queue have teribble behavior for exceptions in a
jobqueue.  They catch all exception and save them for the thread that joins
them.  In practice this leads to deadlocks when exceptions are thrown during
RTA.  We work around this by changing any exception reaching the top level of
an RTA worker into an assertion failure.

Reviewed By: ottoni

Differential Revision: D27435650

fbshipit-source-id: 3169340ae02c4c8bf6b8ad51ffafb1cf1b4ed8fc