serverLazyInit inline and comment naming-table code
Summary:
This is just a small bit of cleanup. I'm inlining the function ServerLazyInit.naming_with_fast, which is only used once. It used to be more complicated, but an earlier diff simplified it to the point where now it's trivial and might as well be incorporated into its caller.

In doing so, I get the added bonus of no longer needing to give it a name with "fast" as a substring. No one knows what Fast even stands for in our codebase.

I added some more comments too.

Reviewed By: CatherineGasnier

Differential Revision: D27414507

fbshipit-source-id: c817b83e9defe3fc9868b097522432ceb06eae85