Simple flow analysis for readonly
Summary:
Before this diff, local variables could only be assigned a single readonlyness. We add basic scoping rules and flow analysis in this diff for all block level constructs. This includes Try, For, Foreach, While and If statements.

We don't do continuation based flow typing here for simplicity's sake(as this will need to be implemented in HackC as well), but instead join at every meeting point at the end of these types of statements. When joining readonlyness, the union type of (Mut | Readonly) is equivalent to the type Readonly, so we can just take the most conservative result in every case.

Because our scoping rules aren't as smart as the typechecker's full flow typing, there are cases where we need to be careful about how we handle join points.
For example:
```
if($b) {
    $x = new Foo();
  } else {
    $x = readonly new Foo();
    return;
  }
```
Since we won't be doing terminality analysis, we won't know that $x is always mutable at the end of the if statement. Thus, if we tried to modify $x after this if statement, we would still error.

Another interesting example:
```
if($b) {
    $x = readonly new Foo();
  } else {
    noreturn_function();
  }
```

Here, if you don't know that noreturn_function() does not return (due to not having typing information), it may seem like $x is not always defined. But we need to conservatively assume that $x is defined after the if statement as a readonly value. Therefore, modifying the properties of $x should fail.

Reviewed By: madgen

Differential Revision: D27059655

fbshipit-source-id: 5ba9dd8c035f7d760004024b0eb2191611b46b25