Readonly closures and converting ConstFun
Summary:
This diff is the first in a stack that implements readonly closures given the syntax agreed upon after the HIP followup:
A readonly function is denoted as:

`(readonly function(): void)`

Readonly closures are equivalent to a regular closure whose `__invoke` method is readonly; this is equivalent to saying that all values the closure captures are captured as readonly. These functions/closures are declared as readonly at declaration time:
```
$captured_value = new Foo();
$x = readonly function() : void {
  // ... captured_value is readonly here
};
$y = readonly () ==> {
  //... captured_value is readonly here
};
```

Note that this is not equivalent to a

`readonly (function() : void)`, which can be obtained by taking a value of a function type and converting it to a readonly value:

```
$captured_value = new Foo();
$x = () ==> {
  $captured_value->prop = 4; // can edit here
};
$f = readonly $x; // $f: readonly (function(): void)
```

A value of type `readonly (function(): void)` cannot be invoked safely, since that function might inherently modify its own contents and the values it captures (which is the opposite of what readonly means)
A value of type `readonly (readonly function(): void)` *can* be called safely, since the function is guaranteed at declaration time to capture its values as readonly.

From a subtyping perspective, a readonly function is a subtype of its regular function counterpart.

This diff only adds the parsing changes for readonly functions, which is relatively straightforward. The only real change besides schema_defintion.ml is in type_parser.rs.

The rest of the stack will address the following:

- readonly parsing
- readonly lowering and declaration of readonly closure types
- lowering `readonly () ==> {}` to create readonly functions
- forbidding `readonly (function() : T)` from being called(while allowing (readonly function(): void)) to be called
- Removing all special stuff about ConstFun we had introduced in https://www.internalfb.com/diff/D26562108 (https://github.com/facebook/hhvm/commit/f96315bdabf429e6433bccc3d51d01c5e9a8c15b) to be replaced with this better syntax

Reviewed By: losvald

Differential Revision: D27244431

fbshipit-source-id: da68e9d5aa777e635d79b344807abd592d3820d9