Collection append
Summary:
This diff adds support for appending to readonly value type containers(i.e. dict, vec and shape).

We don't support varray and darray yet, (will add that in a later time if necessary), but since we're getting rid of those I figured they were lower priority.

If a readonly value is appended to a readonly value type, this action succeeds. We do not support appending a readonly value to a mutable data structure and changing the type of that structure yet; we may be able to build this support later, but as of now it would add a bunch of complexity to the flow typing rules earlier that I haven't thought through.

Reviewed By: madgen

Differential Revision: D27225642

fbshipit-source-id: 8a54e526089565f7ec1850488c68875c0923acf2