Introduce AllocBespokeStructDict
Summary: It turns out that NewBespokeStructDict was broken for the "zero keys" case, but wasn't exercised in my small tests. Fix it by adding a new AllocBespokeStructDict opcode for this case. In the future, we'll optimize the "non-zero keys" case to use this opcode, too, if the number of keys is small enough.

Reviewed By: colavitam

Differential Revision: D27448382

fbshipit-source-id: ea59507cbb495c8dd0bbb8fd0c0aef069658f22b