Jumpstart support for struct layouts
Summary:
I started out making this change in a more complicated way - adding a "serialize" and "deserialize" method to bespoke::Layout - but that didn't solve the problem of dispatch.

I think for now it's simpler to just filter to structs in the existing serialization code.

Now that we have jumpstart support, we can enable struct layouts by default and see how they perform.

Reviewed By: colavitam

Differential Revision: D27445404

fbshipit-source-id: cdfb82f80c7eb81eefd2439c32115f5cb49faf02