Replace use of isdigit in hex2bin with explicit 0-9 check
Summary:
According to the comment, `isdigit` is locale-sensitive. I don't know how true this is - I don't have access to the C standard, and I couldn't figure out what the glibc implementation does.

The implementation doesn't work at all if values outside 0-9 are considered digits, and it's easy enough to check that range explicitly, so let's do that and not worry about locales altogether.

Reviewed By: jano

Differential Revision: D27447229

fbshipit-source-id: 29622eef427487d37c60f38bdb65e220ea65a8fb