Ensure TransLocMaker rollback returns a valid TransRange
Summary:
I noticed we were producing invalid trans ranges in certain overflow scenarios.
This could arise if a TransLocMaker failed to reserve its dword in cold or
frozen, and that triggered a rollback.  During the rollback we would produce a
folly::Range with negative size (since we exclude the reserved dword from the
range).  This would cause assertion failures when determining the ranges size.

Reviewed By: ricklavoie

Differential Revision: D27466921

fbshipit-source-id: 015d0a6620b27ad397840e9937d06765cd769c92