Use FP depth index to identify frames in alias classes
Summary:
Instead of using frame's FPRelOffset, use its depth index to identify it in
alias classes.

This is better, as it allows aliasing of frames that are alive at different
times, but happen to be at different stack depths. This frame aliasing saves
alias location slots, allowing us to track more things.

It also kills the last usage of FPRelOffset, which is an abstraction that is
being removed and is confusing when used across frames (relative to which FP?
the one on the heap?).

Reviewed By: ricklavoie

Differential Revision: D27443816

fbshipit-source-id: ae2fa346eca5345085f540f48b4a2c29654fec72