Make all runtime iteration NoInc
Summary:
IterateV and IterateKV do some refcount operations on the array. It's not real protection, though - they don't inc-ref any of the keys and values, for example (and neither should they; that would be expensive).

Eliminate refcounting from this API completely. If the caller is mutating arrays or objects via this API, it's their responsibility to either a) do the refcounting or b) use the Array version (ArrayIter) instead, which is safe, like all Variant APIs.

The two IterateProp... variants are definitely safe, as only the NoInc versions are ever used.

Reviewed By: ricklavoie

Differential Revision: D27491874

fbshipit-source-id: 085adfb0c9926e0f51762b3e28fb0f2f2074317c