Emit parser fatal for prop write on readonly expression
Summary:
Keep track of readonlyness of local expressions such that we can fatal in the following case:

```
$x = (readonly Foo)
$x->inner = 1
```

Line 1 matches to an assignment of `$x` to a readonly expression, so we mark `$x` as a readonly local.

Line 2 matches to a binary expression over equals, where the left is an object getter. If the object instance was previously marked as readonly, then we fatal.

`readonlyness` is evaluated recursively so this handles the deep property case (and the unit test, deep_prop, confirms).

Reviewed By: jamesjwu

Differential Revision: D27063858

fbshipit-source-id: 4617501ab3e43afde2bcd1e1f5e6bf03444aa6f1