Readonly enforcement for QueryM
Summary:
Enforcement of QueryM Readonly for properties
Assume readonly COW Collection types cannot be modified.

```
    $x = <expr>[5];             // <expr> must be mutable => [5] is mutable
(1) $x = <expr>->c;             // <expr>, c must be mutable
    $x = readonly <expr>[5];    // Any for rhs, $x must be readonly
    $x = readonly <expr>->c;    // Any for rhs, $x must be readonly
```

(1) is the check we need for QueryM;
These are all property accesses so we only need to cover that path for this bytecode.

I tried to be conservative when threading more information through to only the places that were necessary for this check.

Reviewed By: ricklavoie

Differential Revision: D27246639

fbshipit-source-id: bebbe80eb86f4f88a8719ef6103f90c09eb332c2