Don't restart on large rebase when a precomputed saved-state is being used
Summary:
Completely disable `ServerRevisionTracker` when using a precomputed saved-state. We cannot rely on the monitor restarting hh_server, because restarting hh_server when using a precomputed saved-state is completely wrong (due to files being able to change after initial init)

Also refactor a bit, so that it is immediately clear from reading the functions what is global state and what is not.

Reviewed By: ljw1004

Differential Revision: D27505860

fbshipit-source-id: eb010de159b37dc941f4fd3533b816d4841da39c