Make case sensitivity collide between static and regular methods
Summary:
When it comes to case sensitivity, we've actually decided static methods and regular method should collide in naming, contrary to my understanding before. We fix this in the diff and update the error messages to be more clear with this change.

So far we detect 3 errors in www where a method and inherited static method differ only in casing.

Reviewed By: Wilfred

Differential Revision: D27506228

fbshipit-source-id: 5f658c0544c63c31c313ab791f74c3c7345a9c1d