Naming_provider.get_x_canon_name should lower-case its argument
Summary:
The function Naming_provider.get_type_canon_name is supposed to find the true capitalization of a name, if you give it a possibly-incorrect capitalization.

Bizarrely, it was expecting callers to turn that name into lowercase before calling it. If they failed to do so, then it gave incorrect results (under sharedmem backend) depending on whether it happened to have cached a previous lookup or not.

This diff changes it to have a more normal API, where it will find the true capitalization of a name no matter what case you give as argument, and it will cache uniformly in all cases.

This change in behavior is only manifest in one test, "naming_table", which was the only place in our codebase which failed to turn into lowercase before calling it. Everyone else in our production codebase did the right thing. So this diff will have no behavioral impact.

Differential Revision: D27493678

fbshipit-source-id: fb45d430de3552b73040725042752f3bfa13ee23