Naming_provider.get_x_canon_name cleanup
Summary:
The function Naming_provider.get_fun_canon_name was written as though it took an unspecificed-capitalization parameter "name", and then constructed a local variable "canon_name_key" which was the lowercase version of it.

That wasn't true. As shown in the previous diff, every caller already turned the argument into lowercase before invoking this function, and indeed they had to otherwise they'd get inconsistent behavior with caching.

The previous diff made it so that callers no longer had to turn it into lowercase; the "turning into lowercase" is done inside the function.

This revealed that the different between "name" and "canon_name_key" was a lie. They were both meant to be lowercase. So this diff cleans it up.

{F565977364}

Why now? I'm working to make symbol-name be an abstract type, not a string. I'm making "SymbolName.t" a different abstract type from "CaseInsensitiveComparableSymbolName.t". This made me pay closer attention to these functions and tests. That's how I discovered the oddity.

Differential Revision: D27496573

fbshipit-source-id: 4ae1bdb14db892582373c09241cc8d9507349cb7