Add test for Enum Class jit deserialization bug
Summary: Adds a new intrinsic attribute `__JitSerdesPriority` (which can only be used when intrinsics are enabled), which causes marked classes to be stored in first in the jumpstart protocol (and in order starting with the lowest "priority"). The new test fails without D27574956 and exposes a bug in the way that we had been conflating enum and enum class types.

Reviewed By: jano

Differential Revision: D27574955

fbshipit-source-id: f4f594053948516407c559a698d16e7c6687d1b0