Lazily guard locals and stack locations in the tracelet selector
Summary:
This diff reworks the tracelet selector's mechanism to compute the guards
needed.  Previously, guards were eagerly inserted for all live VM locations
before processing any bytecode instruction, and then guard relaxation would
eliminate the unnecessary guards.  Although simple, that approach resulted in an
unnecessarily large number of locals and stack locations being tracked in
FrameState, leading to significant JIT overheads for forming regions in
functions with too many locals and/or stack slots used.  This was particularly
wasteful because the tracelet selector has a limit on the number of locations
that it will guard (currently 5) before breaking the region.

This diff switches to a lazy guarding approach to reduce the JIT overhead in
such cases.  With lazy guarding, new guards are added on demand, right before
the VM location is accessed, thus drastically reducing the number locations that
need to be tracked in FrameState for tracelet selection.  To benefit from the
fact that not all locations need to be tracked in FrameState, the vectors of
locals and stack slots in FrameState were replaced with hash maps to allow a
more efficient sparse representation.

This diff also adds a runtime option to allow eagerly guarding regions up to a
specified number of live VM locations.  This limit is set to 0 by default
(meaning, always lazy guarding), but it can be used to partially or completely
disable lazy guarding.

Reviewed By: paulbiss

Differential Revision: D27545816

fbshipit-source-id: c4ac04d75fd3396b0369981afae200f402a26777