Fix refinement when class isn't direct ancestor
Summary:
During refinement tests `is C` and `as C`, we extend the type parameter environment with bounds on type parameters that *necessarily hold* if the test succeeds at runtime. For example, if `$x : C<T>` and we do `$x is D` for `class D extends C<E>`, then we can deduce (if `C` is invariant) that `T = E`. We do this by assuming `t <: u` where `t` is the runtime type in `$x is t`, and `u` is the static type of `$x`, and deducing what constraints *must* hold if this subtype statement is true.

But sometimes we do a test for which `t` is not actually a subtype of `u`. For example, we might do `$x is J` for `$x:I` where `I` and `J` are unrelated interfaces. Unfortunately the logic that determines what constraints to add to type parameters given a (in this case) false assumption `J <: I` will mark the type parameter environment "inconsistent". This has consequences in code downstream of the type test - other constraint checks will be ignored (see new test in this diff).

A simple fix is to avoid the `add_constraint` logic in the case that types are not in a subtype relationship, and that's what is done in this diff.

Future improvement would be instead to consider the consequences that follow from `t & u` being non-empty, which is really what we are witnessing with the type test.

Reviewed By: vsiles

Differential Revision: D27430079

fbshipit-source-id: e669549dab18f080e61199dc20de9c609445ecba