Fix stack offsets when binding resume address of resumables
Summary:
We can't use `spOffBCFromFP()` when binding resume address while jitting
non-resumed code, as stack in non-resumed code is relative to FP, while stack
in resumed code is relative to the actual stack base.

This bug was not causing any issues, as service requests were ignoring stack
offsets and relied on synced vmsp(), but the next diff is breaking that
assumption and does not work without this diff.

This will actually unblock a diff that replaces with FPInvOffset with
SBInvOffset, which is always relative to the stack base, which will ironically
make the original code correct again.

Reviewed By: ricklavoie

Differential Revision: D27581356

fbshipit-source-id: 4748b70fdc6dd931986bcefb252a8443a391365c