Runtime integration for struct arrays
Summary:
Adds support for StructDicts originating from the runtime, using the same logging and layout selection machinery as arrays with Hack creation sites. A C++ array creation site provides a stable identifier in the form of a string that's used to look up the associated RuntimeStruct. Additionally, it provides a mapping between integer indices and keys it intends to set in the struct. It can then use a StructDictInit initializer to create an array with that RuntimeStruct. The StructDictInit will produce logging arrays (properly sampled) during profiling, and generate either vanilla arrays or a specific StructDict after layout selection.

A few notes:
1) Currently the notion of a runtime location is built into RuntimeStruct. I plan to separate this out in the future when we want to support other layout families originating from the runtime.
2) A caller index -> slot translation table is provided so that layout selection is free to reorder/remove keys as desired.
3) StructDictInit's API is not as comprehensive as DArrayInit, but I will expand it as neccesary.

Differential Revision: D27592822

fbshipit-source-id: cd6afe8205f2f6bd0b08611ead562bd59449a6de