Refactor runtime tests
Summary:
Before flipping on experimental production, I want to make sure that we have sufficient coverage of tests for the runtime as well.

This diff introduces the `expression_tree.inc` file which implements a `Code` visitor that transforms the given AST into a string representation. This file can be `require`d by other tests to have easy access to `print_et` which prints a `Code` Expression Tree to verify that the runtime is doing what we expect.

Reviewed By: Wilfred

Differential Revision: D27542070

fbshipit-source-id: fd8ec81c14bc9414e55958c4a7940d954dbdd057