Improve type checking for (Key)ValCollections and Pairs
Summary:
Switch `compute_supertype` to use `check_expected_ty` (which is what `new` and lambdas use) instead of bare `sub_type` for the expected type. This makes it a bit more flexible wrt `dynamic`. This is important for sound dynamic, but should be an improvement in general. In particular, this supports
```
function f() : vec<dynamic> {
  return vec[1];
}
```
by allowing the expected type check for the vec's elements to know that `1 <: dynamic`.

Reviewed By: CatherineGasnier

Differential Revision: D27620812

fbshipit-source-id: d3ddbd65d5595d7cf3ca7ef28e92f77df92d15d2