Remove fallback_expect_extension from run_idempotence_tests
Summary: I'm working on making `verify.py` create a diff for tests with fallbacks and delete unnecessary exp files. The idempotence test is only run for hackfmt; a fallback extension is not specified in the test target and it doesn't make sense in context. This gets rid of the propagated argument.

Reviewed By: Wilfred

Differential Revision: D27640412

fbshipit-source-id: 4b804f22885aa182179ca10d537d94267e5846bc